+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to a section with the version name.
 
 * Track which `Sales Invoice Additional Fields` is latest in case of multiple submissions for the same invoice  due to
   rejection
+* Limit fixing rejection to the latest sales invoice additional fields document
 
 ## 0.20.2