+++ CHANGELOG.md
@@ -8,6 +8,8 @@ to a section with the version name.
 
 ## Unreleased Changes
 
+## 0.21.0
+
 * Track which `Sales Invoice Additional Fields` is latest in case of multiple submissions for the same invoice  due to
   rejection
 * Limit fixing rejection to the latest sales invoice additional fields document